# Use native diff log for git