--- a/lib/vcs/git.js
+++ b/lib/vcs/git.js
@@ -1,19 +1,20 @@
-var exec = require('child_process').exec;
+var exec = require('child_process').exec,
+	LogParser = require('../parser/gitlog'),
+	parser = new LogParser();
+	util = require('util');
 
 function Git() {
 
 }
 
-Git.prototype.getLog = function (branch, minRevision, callback) {
-	exec("git log --no-merges '" + branch+"'", {maxBuffer: 1024 * 10000}, function (err, stdout, stderr) {
+Git.prototype.getDiffLog = function (fromBranch, toBranch, callback) {
+	var command = util.format("git log --no-merges '%s'..'%s'", fromBranch, toBranch);
+
+	exec(command, {maxBuffer: 1024 * 1024 * 10}, function (err, stdout, stderr) {
 		if (err) throw err;
 
-		callback(stdout);
+		callback(parser.parse(stdout));
 	});
-};
-
-Git.prototype.getMinVersion = function (oldBranch, newBranch, callback) {
-	callback(0);
 };
 
 module.exports = Git;