--- a/lib/vcs/svn.js
+++ b/lib/vcs/svn.js
@@ -2,11 +2,33 @@ var exec = require('child_process').exec,
 	ArrayHelper = require('../util/arrayhelper'),
 	arrayHelper = new ArrayHelper(),
 	config = require('../../config'),
+	Builder = require('../builder'),
 	util = require('util');
 
 function Svn() {
 
 }
+
+Svn.prototype.getDiffLog = function (fromBranch, toBranch, callback) {
+	var builder = new Builder();
+	var count = 0;
+
+	this.getMinVersion(fromBranch, toBranch, function (minRevision) {
+		this.getLog(fromBranch, minRevision, function (log) {
+			builder.setOldLog(log);
+			if (count++) {
+				callback(builder.getDiff());
+			}
+		});
+
+		this.getLog(toBranch, minRevision, function (log) {
+			builder.setNewLog(log);
+			if (count++) {
+				callback(builder.getDiff());
+			}
+		});
+	}.bind(this));
+};
 
 Svn.prototype.getLog = function (branch, minRevision, callback) {
 	this._executeSvnCommand(util.format("log -g -r %s:HEAD %s%s", minRevision, config.vcs.path, branch),  function (err, stdout, stderr) {
@@ -17,14 +39,13 @@ Svn.prototype.getLog = function (branch, minRevision, callback) {
 };
 
 Svn.prototype.getMinVersion = function (oldBranch, newBranch, callback) {
-	var self = this;
 	this.getRevisionsForBranch(oldBranch, function (oldRevisions) {
-		self.getRevisionsForBranch(newBranch, function (newRevisions) {
+		this.getRevisionsForBranch(newBranch, function (newRevisions) {
 			var revisionDiff = arrayHelper.diff(newRevisions, oldRevisions);
 
 			callback(Math.min.apply(null, revisionDiff) - 1);
 		});
-	});
+	}.bind(this));
 };
 
 Svn.prototype.getRevisionsForBranch = function (branch, callback) {
@@ -65,7 +86,7 @@ Svn.prototype._executeSvnCommand = function(command, callback) {
 		command_string += util.format(" --username=%s --password=%s", config.vcs.user, config.vcs.password);
 	}
 
-	exec(command_string, {maxBuffer: 1024 * 10000}, callback);
+	exec(command_string, {maxBuffer: 1024 * 1024 * 100}, callback);
 };
 
 module.exports = Svn;