Instead of diff'ing huge logs use the native diff log for git. This is faster and requires less memory while buffering the log outputs.

```
time node ../changelog-builder/index.js release/1.60 release/1.61

Old: 3,96s user 0,33s system 84% cpu 5,081 total
New: 0,24s user 0,05s system 14% cpu 1,997 total
```
There is no significant change for the SVN behavior.